+++ src/app/dynamic-form/dynamic-form.component.ts
@@ -1,4 +1,4 @@
-import { Input, OnInit, ComponentFactoryResolver, ViewContainerRef, Directive, Inject } from '@angular/core';
+import { Input, OnInit, ComponentFactoryResolver, ViewContainerRef, Directive, Inject, AfterViewInit } from '@angular/core';
 import { FormGroup } from '@angular/forms';
 import { FormConfig, LayoutDictionary, LAYOUTS_TOKEN, Layout } from '../../types';
 
@@ -6,10 +6,10 @@ import { FormConfig, LayoutDictionary, LAYOUTS_TOKEN, Layout } from '../../types
     exportAs: 'dynamicForm',
     selector: 'dynamic-form'
 })
-export class DynamicFormDirective implements OnInit {
+export class DynamicFormDirective implements OnInit, AfterViewInit {
     @Input() formConfig: FormConfig;
     @Input() model: any;
-    @Input() lookups: object;
+    @Input() readOnly: boolean;
 
     public group: FormGroup;
     get changes() { return this.group.valueChanges; }
@@ -33,6 +33,12 @@ export class DynamicFormDirective implements OnInit {
         component.instance.group = this.group;
         component.instance.formConfig = this.formConfig;
         component.instance.model = this.model;
+    }
+
+    ngAfterViewInit() {
+        if (this.readOnly) {
+            this.group.disable();
+        }
     }
 
 }