+++ src/app/fields/form-date/form-date.component.ts
@@ -1,8 +1,15 @@
 import { Component } from '@angular/core';
 import { BaseFieldComponent } from '../base-field';
+import { NgbDateAdapter } from '@ng-bootstrap/ng-bootstrap';
+import { DatePickerAdapter } from './data-adapter/date-picker.adapter';
 
 @Component({
     selector: 'form-date',
-    template: require('./form-date.component.html')
+    template: require('./form-date.component.html'),
+    providers: [
+        {
+            provide: NgbDateAdapter,
+            useClass: DatePickerAdapter
+        }]
 })
-export class FormDateComponent extends BaseFieldComponent {}
+export class FormDateComponent extends BaseFieldComponent { }